deploy_v2/services/socket-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
PyJWT==2.8.0
httpx==0.27.0
websockets==12.0
redis[hiredis]==5.0.7
